feat: Refactor query handling and introduce new Query decorator
- Removed the `Query` decorator implementation from `crud_repository_implementation_service.py` and moved it to a new `query.py` file for better organization.
- Updated the `ConditionNotation` class to inherit from `str` and `Enum` for improved functionality.
- Cleaned up imports in `crud_repository_implementation_service.py` and added necessary imports in `query.py`.
- Enhanced the `Query` decorator with detailed docstrings and improved type handling for SQL query generation.
