+++ py_spring_model/__init__.py
@@ -2,4 +2,5 @@ from py_spring_model.core.model import PySpringModel, Field
 from py_spring_model.py_spring_model_provider import provide_py_spring_model
 from py_spring_model.repository.crud_repository import CrudRepository
 from py_spring_model.repository.repository_base import RepositoryBase
-from py_spring_model.py_spring_model_rest.service.curd_repository_implementation_service.crud_repository_implementation_service import Query, SkipAutoImplmentation
+from py_spring_model.py_spring_model_rest.service.curd_repository_implementation_service.crud_repository_implementation_service import SkipAutoImplmentation
+from py_spring_model.py_spring_model_rest.service.query_service.query import Query
